add relay client seperation
