+++ pkg/pokt/pokt_v0/basic_client.go
@@ -1,13 +1,10 @@
 package pokt_v0
 
 import (
-	"encoding/hex"
 	"github.com/pquerna/ffjson/ffjson"
 	"github.com/valyala/fasthttp"
-	"math/rand"
 	"os-gateway/pkg/common"
 	"os-gateway/pkg/pokt/pokt_v0/models"
-	"slices"
 	"time"
 )
 
@@ -77,7 +74,7 @@ func (r BasicClient) SendRelay(req *models.SendRelayRequest) (*models.SendRelayR
 	}
 
 	// Get the preferred selected node, or chose a random one.
-	node, err := r.getNodeFromRequest(session, req.SelectedNodePubKey)
+	node, err := getNodeFromRequest(session, req.SelectedNodePubKey)
 
 	if err != nil {
 		return nil, err
@@ -87,14 +84,14 @@ func (r BasicClient) SendRelay(req *models.SendRelayRequest) (*models.SendRelayR
 
 	relayMetadata := &models.RelayMeta{BlockHeight: currentSessionHeight}
 
-	relayProof := r.generateRelayProof(req.Chain, currentSessionHeight, node.PublicKey, relayMetadata, req.Payload, req.Signer)
+	relayProof := generateRelayProof(req.Chain, currentSessionHeight, node.PublicKey, relayMetadata, req.Payload, req.Signer)
 
 	// Relay created, generating a request to the servicer
 	var sessionResponse models.SendRelayResponse
 	err = r.makeRequest(endpointSendRelay, "POST", &models.Relay{
 		Payload:    req.Payload,
 		Metadata:   relayMetadata,
-		RelayProof: &relayProof,
+		RelayProof: relayProof,
 	}, &sessionResponse, &node.ServiceUrl)
 
 	if err != nil {
@@ -161,125 +158,4 @@ func (r BasicClient) makeRequest(endpoint string, method string, requestData any
 		return pocketError
 	}
 	return ffjson.Unmarshal(response.Body(), responseModel)
-}
-
-// generateRelayProof generates a relay proof.
-// Parameters:
-//   - chainId: Blockchain ID.
-//   - sessionHeight: Session block height.
-//   - servicerPubKey: Servicer public key.
-//   - requestMetadata: Request metadata.
-//   - account: Ed25519 account used for signing.
-//
-// Returns:
-//   - models.RelayProof: Generated relay proof.
-func (r BasicClient) generateRelayProof(chainId string, sessionHeight uint, servicerPubKey string, relayMetadata *models.RelayMeta, reqPayload *models.Payload, account *models.Ed25519Account) models.RelayProof {
-	entropy := uint64(rand.Int63())
-	aat := account.GetAAT()
-
-	requestMetadata := models.RequestHashPayload{
-		Metadata: relayMetadata,
-		Payload:  reqPayload,
-	}
-
-	requestHash := requestMetadata.Hash()
-
-	unsignedAAT := &models.AAT{
-		Version:      aat.Version,
-		AppPubKey:    aat.AppPubKey,
-		ClientPubKey: aat.ClientPubKey,
-		Signature:    "",
-	}
-
-	proofObj := &models.RelayProofHashPayload{
-		RequestHash:        requestHash,
-		Entropy:            entropy,
-		SessionBlockHeight: sessionHeight,
-		ServicerPubKey:     servicerPubKey,
-		Blockchain:         chainId,
-		Signature:          "",
-		UnsignedAAT:        unsignedAAT.Hash(),
-	}
-
-	hashedPayload := common.Sha3_256Hash(proofObj)
-	hashSignature := hex.EncodeToString(account.Sign(hashedPayload))
-	return models.RelayProof{
-		RequestHash:        requestHash,
-		Entropy:            entropy,
-		SessionBlockHeight: sessionHeight,
-		ServicerPubKey:     servicerPubKey,
-		Blockchain:         chainId,
-		AAT:                aat,
-		Signature:          hashSignature,
-	}
-}
-
-// getSessionFromRequest obtains a session from a relay request.
-// Parameters:
-//   - req: SendRelayRequest instance containing the relay request parameters.
-//
-// Returns:
-//   - (*GetSessionResponse): Session response.
-//   - (error): Error, if any.
-func GetSessionFromRequest(pocketService PocketService, req *models.SendRelayRequest) (*models.Session, error) {
-	if req.Session != nil {
-		return req.Session, nil
-	}
-	sessionResp, err := pocketService.GetSession(&models.GetSessionRequest{
-		AppPubKey: req.Signer.PublicKey,
-		Chain:     req.Chain,
-	})
-	if err != nil {
-		return nil, err
-	}
-	return sessionResp.Session, nil
-}
-
-// getNodeFromRequest obtains a node from a relay request.
-// Parameters:
-//   - req: SendRelayRequest instance containing the relay request parameters.
-//
-// Returns:
-//   - (*models.Node): Node instance.
-//   - (error): Error, if any.
-func (r BasicClient) getNodeFromRequest(session *models.Session, selectedNodePubKey string) (*models.Node, error) {
-	if selectedNodePubKey == "" {
-		return getRandomNodeOrError(session.Nodes, models.ErrSessionHasZeroNodes)
-	}
-	return findNodeOrError(session.Nodes, selectedNodePubKey, models.ErrNodeNotFound)
-}
-
-// getRandomNodeOrError gets a random node or returns an error if the node list is empty.
-// Parameters:
-//   - nodes: List of nodes.
-//   - err: Error to be returned if the node list is empty.
-//
-// Returns:
-//   - (*models.Node): Random node.
-//   - (error): Error, if any.
-func getRandomNodeOrError(nodes []*models.Node, err error) (*models.Node, error) {
-	node := common.GetRandomElement(nodes)
-	if node == nil {
-		return nil, err
-	}
-	return node, nil
-}
-
-// findNodeOrError finds a node by public key or returns an error if the node is not found.
-// Parameters:
-//   - nodes: List of nodes.
-//   - pubKey: Public key of the node to find.
-//   - err: Error to be returned if the node is not found.
-//
-// Returns:
-//   - (*models.Node): Found node.
-//   - (error): Error, if any.
-func findNodeOrError(nodes []*models.Node, pubKey string, err error) (*models.Node, error) {
-	idx := slices.IndexFunc(nodes, func(node *models.Node) bool {
-		return node.PublicKey == pubKey
-	})
-	if idx == -1 {
-		return nil, err
-	}
-	return nodes[idx], nil
 }
